add feature to allow user to upload amount and description from paystub to db
