+++ .idea/modules.xml
@@ -2,8 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/TipTracker.iml" filepath="$PROJECT_DIR$/TipTracker.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/tiptracker.iml" filepath="$PROJECT_DIR$/tiptracker.iml" />
     </modules>
   </component>
 </project>
